SQLite Test

+++ Biwen.EFCore.SoftDelete.TestConsole/Program.cs
@@ -4,12 +4,12 @@ using Biwen.EFCore.SoftDelete.TestConsole;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Biwen.EFCore.SoftDelete.TestConsole.Domains;
-using Biwen.EFCore.SoftDelete;
 
 var serviceProvider = new ServiceCollection()
     .AddDbContext<TestDbContext>(options =>
     {
-        options.UseInMemoryDatabase("test");
+        //options.UseInMemoryDatabase("test");
+        options.UseSqlite("Data Source=../../../test.db");
     })
     //.Decorate<TestDbContext>((ctx) => new SoftDeleteDecorator<TestDbContext>(ctx).DbContext)
     .BuildServiceProvider();
@@ -18,22 +18,33 @@ var serviceProvider = new ServiceCollection()
 serviceProvider.GetRequiredService<TestDbContext>().Database.EnsureCreated();
 
 using var sp = serviceProvider.CreateScope();
-var db = sp.ServiceProvider.GetRequiredService<TestDbContext>();
+var db = sp.ServiceProvider.GetRequiredService<TestDbContext>()!;
+
+
+//delete all rows
+db.Database.ExecuteSqlRaw("DELETE FROM blogs");
+
+List<int> ids = new();
 
 int i = 1;
 while (i <= 5)
 {
-    db.Blogs.Add(new Blog
+    var blog = new Blog
     {
         Title = $"test {Guid.NewGuid()}",
         Content = "test content",
         Tags = "test",
         AuthorId = 1,
-    });
+    };
+
+
+    db.Blogs.Add(blog);
     i++;
+    db.SaveChanges();
+
+    ids.Add(blog.Id);
 }
 
-db.SaveChanges();
 
 var blogs = db.Blogs.ToList();
 foreach (var blog in blogs)
@@ -44,20 +55,23 @@ foreach (var blog in blogs)
 Console.WriteLine("-----↑----初始化的数据-----↑--------");
 
 //Delete 1 模拟软删除
-var blog1 = db.Blogs.FirstOrDefault(x => x.Id == 1);
+var blog1 = db.Blogs.FirstOrDefault(x => x.Id == ids[0]);
 db.Remove(blog1!);
 db.SaveChanges();
 
 //Delete 2 模拟软删除
-var blog2 = db.Blogs.FirstOrDefault(x => x.Id == 2);
+var blog2 = db.Blogs.FirstOrDefault(x => x.Id == ids[1]);
 db.Remove(blog2!);
 db.SaveChanges();
 
 
 //Delete 3 模拟强制删除
-var blog3 = db.Blogs.FirstOrDefault(x => x.Id == 3);
-db.Remove(blog3!, true);
-db.SaveChanges();
+var blog3 = db.Blogs.FirstOrDefault(x => x.Id == ids[2]);
+if (blog3 != null)
+{
+    db.Remove(blog3!, true);
+    db.SaveChanges();
+}
 
 
 var blogs2 = db.Blogs.ToList();
